[front] refactor: 커스텀 훅 구조 리팩토링
