+++ front/src/hooks/useInterval.ts
@@ -1,0 +1,17 @@
+// https://overreacted.io/making-setinterval-declarative-with-react-hooks/
+import { useEffect, useRef } from 'react';
+
+export const useInterval = (callback: Function, delay: number) => {
+  const savedCallback = useRef<Function>()
+  useEffect(() => {
+    savedCallback.current = callback
+  }, [callback])
+  useEffect(() => {
+    const handler = (...args: any) => savedCallback.current?.(...args)
+
+    if (delay !== null) {
+      const id = setInterval(handler, delay)
+      return () => clearInterval(id)
+    }
+  }, [delay])
+}
